requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
geographiclib==2.0
geopy==2.4.1
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg==3.1.18
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
whitenoise==6.6.0
